Added spring profile "springJdbc". Turning on this profile replaces all springDataJpa repositories with custom implementations made with Spring Jdbc + QueryDSL.
Application functionality stay same as with "springDataJpa" profile, but only used in this app repositories methods have realization (others will throw operation not supported exception).

Only simple QueryDSL queries are supported (for example subqueries are not supported and so on).

This PR should be considered only as illustration of inconvenience of Spring Jdbc usage for classical CRUD API